+++ app/components/deposit-flow/deposit.tsx
@@ -2,6 +2,7 @@ import { ICON_SIZE } from "~/lib/constants";
 import type { SwapRow, SwapRowMarketDatum } from "~/types/global";
 import { useEffect, useRef, useState } from "react";
 import type { JsonRpcSigner } from "@ethersproject/providers";
+import { useValidInput } from "~/hooks/use-valid-input";
 import toast from "react-hot-toast";
 import Max from "~/components/max";
 
@@ -29,12 +30,12 @@ export default function Deposit({
   borrowLimitUsed,
   walletBalance,
 }: Props) {
-  let [isValid, setIsValid] = useState<boolean>(false);
   let [isEnabled, setIsEnabled] = useState<boolean>(true);
   let [isEnabling, setIsEnabling] = useState<boolean>(false);
   let [isDepositing, setIsDepositing] = useState<boolean>(false);
   let [value, setValue] = useState<string>("0");
   let inputEl = useRef<HTMLInputElement>(null);
+  let isValid = useValidInput(value, 0, walletBalance);
 
   useEffect(() => {
     if (!signer) {
@@ -48,27 +49,6 @@ export default function Deposit({
       }
     );
   }, [signer, row.cToken, row.token]);
-
-  useEffect(() => {
-    try {
-      if (isNaN(parseFloat(value))) {
-        throw "NaN";
-      }
-
-      let v: number = parseFloat(value);
-
-      let floor: number = 0;
-      let ciel: number = walletBalance;
-      if (v > floor && v < ciel) {
-        setIsValid(true);
-      } else {
-        setIsValid(false);
-      }
-    } catch (e) {
-      console.error(`Invalid input`, e);
-      setIsValid(false);
-    }
-  }, [value, walletBalance]);
 
   useEffect(() => {
     inputEl && inputEl.current && inputEl.current.select();
